Add email code (still to be moved to the right place once it exists)

+++ app/src/main/java/at/ac/tuwien/mns/mnsgeolocation/dto/LatLng.kt
@@ -4,6 +4,6 @@ package at.ac.tuwien.mns.mnsgeolocation.dto
  * Created by johannesvass on 03.01.18.
  */
 class LatLng {
-    var lat: Float? = null
-    var lng: Float? = null
+    var lat: Double? = null
+    var lng: Double? = null
 }
